# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Heiba private video streaming platform API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token in the Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: heiba_session
      description: Session token in an httpOnly cookie
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    VideoStatus:
      type: string
      enum: [PROCESSING, PENDING_REVIEW, APPROVED, REJECTED, PRIVATE, ARCHIVED, FAILED]
    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/VideoStatus"
        durationSeconds:
          type: number
          nullable: true
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          nullable: true
        originalFileName:
          type: string
          nullable: true
        storageProvider:
          type: string
          nullable: true
        uploadedBy:
          type: integer
        categoryIds:
          type: array
          items:
            type: integer
        groupIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - tags
        - status
        - uploadedBy
        - categoryIds
        - groupIds
        - createdAt
        - updatedAt
    VideoListResult:
      type: object
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        total:
          type: integer
      required:
        - videos
        - total
    CreateVideoInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: string
          maxLength: 5000
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        categoryIds:
          type: array
          items:
            type: integer
        groupIds:
          type: array
          items:
            type: integer
      required:
        - title
    UpdateVideoInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: string
          maxLength: 5000
          nullable: true
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        categoryIds:
          type: array
          items:
            type: integer
        groupIds:
          type: array
          items:
            type: integer
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List videos for management (OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/VideoStatus"
        - name: groupId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Videos list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoListResult"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Create a video entry (OWNER/ADMIN)
      description: Creates metadata; upload the binary via POST /videos/{id}/file. New videos enter PROCESSING until the upload completes, then PENDING_REVIEW.
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoInput"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a video for management (OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateVideo
      tags: [videos]
      summary: Update video metadata and assignments (OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoInput"
      responses:
        "200":
          description: Video updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video and its stored binary (OWNER/ADMIN)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Video deleted
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /videos/{id}/file:
    post:
      operationId: uploadVideoFile
      tags: [videos]
      summary: Upload the video binary (OWNER/ADMIN)
      description: Accepts a single multipart "file" field (mp4/webm/mkv/mov). The video moves to PENDING_REVIEW once stored.
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload stored; video is now PENDING_REVIEW
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Missing file or unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...