# lib/api-zod/src/generated/types/updateVideoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Heiba private video streaming platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateVideoInput {
  /**
     * @minLength 1
     * @maxLength 300
     */
  title?: string;
  /** @maxLength 5000 */
  description?: string | null;
  /**
     * @maxItems 20
     * @items.maxLength 50
     */
  tags?: string[];
  categoryIds?: number[];
  groupIds?: number[];
}