# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Heiba private video streaming platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoStatus } from './videoStatus';

export interface Video {
  id: number;
  title: string;
  description?: string | null;
  tags: string[];
  status: VideoStatus;
  durationSeconds?: number | null;
  mimeType?: string | null;
  sizeBytes?: number | null;
  originalFileName?: string | null;
  storageProvider?: string | null;
  uploadedBy: number;
  categoryIds: number[];
  groupIds: number[];
  createdAt: Date;
  updatedAt: Date;
}