# lib/api-zod/src/generated/types/videoStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Heiba private video streaming platform API
 * OpenAPI spec version: 0.1.0
 */

export type VideoStatus = typeof VideoStatus[keyof typeof VideoStatus];


export const VideoStatus = {
  PROCESSING: 'PROCESSING',
  PENDING_REVIEW: 'PENDING_REVIEW',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  PRIVATE: 'PRIVATE',
  ARCHIVED: 'ARCHIVED',
  FAILED: 'FAILED',
} as const;